working on hooks

+++ src/colour.c
@@ -4,7 +4,7 @@
 
 #include "../fractol.h"
 
-int		ft_getcolour(int i, int i_max)
+int		ft_getcolour(double i, int i_max)
 {
 	int		colour = (int)((double)i * (double)GREEN / (double)i_max);
